requirements.txt
beautifulsoup4==4.4.1
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
nose==1.3.7
python-dateutil==1.5
pytz==2015.7
requests==2.9.1
stravalib==0.5.0
units==0.6
Werkzeug==0.11.4
wheel==0.26.0
